# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Starter Kit API",
        "description": "API documentation for the Laravel Filament Starter Kit.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Current application environment"
        }
    ],
    "paths": {
...
        "/api/v1/comments": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Daftar komentar",
                "operationId": "31364881058187fc38a42287e9d904a1",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "approved",
                                "pending"
                            ]
                        }
                    },
                    {
                        "name": "commentable_type",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "commentable_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar komentar berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CommentResource"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/posts/{post}/comments": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Daftar komentar untuk post",
                "operationId": "1a290d43d416e9cc2ca0c0576b2d7a0d",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "Slug post",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar komentar berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CommentResource"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Post tidak ditemukan"
                    }
                }
            },
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Kirim komentar pada post",
                "operationId": "46402843838e49c6deed32d553e785c3",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "Slug post",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "content"
                                ],
                                "properties": {
                                    "content": {
                                        "type": "string",
                                        "example": "Komentar yang sangat membantu"
                                    },
                                    "parent_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "guest_name": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "guest_email": {
                                        "type": "string",
                                        "format": "email",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Komentar tersimpan (langsung approve)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommentResource"
                                }
                            }
                        }
                    },
                    "202": {
                        "description": "Komentar tersimpan dan menunggu moderasi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommentResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Post tidak ditemukan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                }
            }
        },
        "/api/v1/comments/{comment}/approve": {
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Approve komentar",
                "operationId": "3a043c32e80e418a541a0ea92ec43afc",
                "parameters": [
                    {
                        "name": "comment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Komentar berhasil di-approve",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommentResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Tidak memiliki izin"
                    },
                    "404": {
                        "description": "Komentar tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/comments/{comment}": {
            "delete": {
                "tags": [
                    "Comments"
                ],
                "summary": "Hapus komentar",
                "operationId": "c14b4d6f903f4a5dd74e90992d006b95",
                "parameters": [
                    {
                        "name": "comment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Komentar dihapus"
                    },
                    "403": {
                        "description": "Tidak memiliki izin"
                    },
                    "404": {
                        "description": "Komentar tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CommentResource": {
                "title": "Comment",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 45
                    },
                    "content": {
                        "type": "string",
                        "example": "Komentar yang sangat membantu"
                    },
                    "is_approved": {
                        "type": "boolean",
                        "example": true
                    },
                    "guest_name": {
                        "type": "string",
                        "nullable": true
                    },
                    "guest_email": {
                        "type": "string",
                        "nullable": true
                    },
                    "parent_id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "metadata": {
                        "type": "object",
                        "nullable": true
                    },
                    "commentable_type": {
                        "type": "string",
                        "example": "App\\\\Models\\\\Post"
                    },
                    "commentable_id": {
                        "type": "integer",
                        "example": 12
                    },
                    "replies_count": {
                        "type": "integer",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserResource"
                    }
                },
                "type": "object"
            },
...
            "UserResource": {
                "title": "User",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Admin User"
                    },
                    "username": {
                        "type": "string",
                        "example": "admin"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "admin@example.com"
                    },
                    "avatar": {
                        "type": "string",
                        "example": "https://...",
                        "nullable": true
                    },
                    "bio": {
                        "type": "string",
                        "nullable": true
                    },
                    "phone": {
                        "type": "string",
                        "nullable": true
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "permissions": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "preferences": {
                        "type": "object",
                        "nullable": true
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "last_login_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}